Synchronous invocations in Entry/Exit proxies.
The goal is to make the simulation advance lazily, without ever
invoking World.run() explicitly. Instead, the simulation will
advance only as much as needed to answer Entry<T>.call() invocations.

This is however incomplete, as Timeline needs to be changed
to allow stopping the simulation without letting it run to
completion (i.e. all events).

This revision also removes the need for asynchronous methods to
return void. Instead, 0/false/null will be returned as appropriate
and should be discarded.

Fixes issue 7.