Extended API to wait on Exit objects too.
The simulaton can now be run until a set of target milestones
is met. A milestone is a transition from simulation to the
driver code. This means either an invocationin a exit object
or a return in an entry object.

This also fixes a bug in which World.runAll(...) might not
really wait for all entries (now, all milestones).