Make all simulation objects Closeable.
The main goal is to make simulation threads die, such that simulation
state can be garbage collected. This is currently done by waking them
up and making them throw a SimulationThreadDeath error, that, hopefully,
is not catched by application code.

Simulation can only be closed while stopped. And, currently, only
the simulation as a whole should be closed. Support for "killing"
processes and hosts by closing them is planned.