Fix a couple of bugs in the parallel simulation kernel.
The first was in the scheduler, where it was possible to advance the
lease without having considered all timelines. This led to exceeding
the fuzzyness limit.

The second was in the timeline where an iterator was being used. This
allowed for out of order execution of events as the iterator skipped
over an event that was inserted just after the one that had been
executed.